refactor: route api in proxy
