+++ proxy/src/http/mod.rs
@@ -1,4 +1,5 @@
 // Copyright 2023 CeresDB Project Authors. Licensed under Apache-2.0.
 
 pub mod prom;
+pub mod route;
 pub mod sql;
