+++ server/src/mysql/builder.rs
@@ -4,12 +4,11 @@ use std::{net::SocketAddr, sync::Arc, time::Duration};
 
 use proxy::Proxy;
 use query_engine::executor::Executor as QueryExecutor;
-use router::RouterRef;
 use snafu::{OptionExt, ResultExt};
 use table_engine::engine::EngineRuntimes;
 
 use crate::mysql::{
-    error::{MissingInstance, MissingRouter, MissingRuntimes, ParseIpAddr, Result},
+    error::{MissingInstance, MissingRuntimes, ParseIpAddr, Result},
     service::MysqlService,
 };
 
@@ -17,7 +16,6 @@ pub struct Builder<Q> {
     config: Config,
     runtimes: Option<Arc<EngineRuntimes>>,
     proxy: Option<Arc<Proxy<Q>>>,
-    router: Option<RouterRef>,
 }
 
 #[derive(Debug)]
@@ -33,7 +31,6 @@ impl<Q> Builder<Q> {
             config,
             runtimes: None,
             proxy: None,
-            router: None,
         }
     }
 
@@ -46,24 +43,18 @@ impl<Q> Builder<Q> {
         self.proxy = Some(proxy);
         self
     }
-
-    pub fn router(mut self, router: RouterRef) -> Self {
-        self.router = Some(router);
-        self
-    }
 }
 
 impl<Q: QueryExecutor + 'static> Builder<Q> {
     pub fn build(self) -> Result<MysqlService<Q>> {
         let runtimes = self.runtimes.context(MissingRuntimes)?;
         let proxy = self.proxy.context(MissingInstance)?;
-        let router = self.router.context(MissingRouter)?;
 
         let addr: SocketAddr = format!("{}:{}", self.config.ip, self.config.port)
             .parse()
             .context(ParseIpAddr { ip: self.config.ip })?;
 
-        let mysql_handler = MysqlService::new(proxy, runtimes, router, addr, self.config.timeout);
+        let mysql_handler = MysqlService::new(proxy, runtimes, addr, self.config.timeout);
         Ok(mysql_handler)
     }
 }
