+++ server/src/mysql/service.rs
@@ -7,7 +7,6 @@ use log::{error, info};
 use opensrv_mysql::AsyncMysqlIntermediary;
 use proxy::Proxy;
 use query_engine::executor::Executor as QueryExecutor;
-use router::RouterRef;
 use table_engine::engine::EngineRuntimes;
 use tokio::sync::oneshot::{self, Receiver, Sender};
 
@@ -16,7 +15,6 @@ use crate::mysql::{error::Result, worker::MysqlWorker};
 pub struct MysqlService<Q> {
     proxy: Arc<Proxy<Q>>,
     runtimes: Arc<EngineRuntimes>,
-    router: RouterRef,
     socket_addr: SocketAddr,
     join_handler: Option<JoinHandle<()>>,
     tx: Option<Sender<()>>,
@@ -27,14 +25,12 @@ impl<Q> MysqlService<Q> {
     pub fn new(
         proxy: Arc<Proxy<Q>>,
         runtimes: Arc<EngineRuntimes>,
-        router: RouterRef,
         socket_addr: SocketAddr,
         timeout: Option<Duration>,
     ) -> MysqlService<Q> {
         Self {
             proxy,
             runtimes,
-            router,
             socket_addr,
             join_handler: None,
             tx: None,
@@ -55,7 +51,6 @@ impl<Q: QueryExecutor + 'static> MysqlService<Q> {
         self.join_handler = Some(rt.default_runtime.spawn(Self::loop_accept(
             self.proxy.clone(),
             self.runtimes.clone(),
-            self.router.clone(),
             self.socket_addr,
             self.timeout,
             rx,
@@ -72,7 +67,6 @@ impl<Q: QueryExecutor + 'static> MysqlService<Q> {
     async fn loop_accept(
         proxy: Arc<Proxy<Q>>,
         runtimes: Arc<EngineRuntimes>,
-        router: RouterRef,
         socket_addr: SocketAddr,
         timeout: Option<Duration>,
         mut rx: Receiver<()>,
@@ -93,11 +87,10 @@ impl<Q: QueryExecutor + 'static> MysqlService<Q> {
                         }
                     };
                     let proxy = proxy.clone();
-                    let router = router.clone();
 
                     let rt = runtimes.read_runtime.clone();
                     rt.spawn(AsyncMysqlIntermediary::run_on(
-                        MysqlWorker::new(proxy, router, timeout),
+                        MysqlWorker::new(proxy,  timeout),
                         stream,
                     ));
                 },
